# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: VR-GeoNusa API
  version: 0.6.0
paths:
...
  /api/v1/learning-sessions:
    post:
      operationId: createLearningSession
      summary: Create a learning session for the authenticated student
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLearningSessionRequest"
      responses:
        "201":
          description: Learning session created; the write token is returned only once
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateLearningSessionResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Respondent code does not belong to the authenticated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: The student already has an active learning session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Invalid input or client-supplied session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/learning-sessions/{id}/panorama-visits:
    post:
      operationId: recordPanoramaVisit
      summary: Record a panorama visit during exploration
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/LearningSessionId"
        - $ref: "#/components/parameters/SessionWriteToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordPanoramaVisitRequest"
      responses:
        "201":
          description: Panorama visit recorded and current panorama updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityEventResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Invalid session write token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Learning session not found or not owned by the student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Learning session is not in the exploration phase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Invalid panorama node or client-supplied visit state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
...
    LearningSessionId:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
    SessionWriteToken:
      name: X-Session-Write-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 64
        maxLength: 64
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum token
  schemas:
    RecordPanoramaVisitRequest:
      type: object
      additionalProperties: false
      required: [panorama_node_id]
      properties:
        panorama_node_id:
          type: integer
    ActivityEventResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/PanoramaVisitedEvent"
    PanoramaVisitedEvent:
      type: object
      required: [id, type, panorama_node, visited_at]
      properties:
        id:
          type: integer
        type:
          type: string
          const: panorama_visited
        panorama_node:
          $ref: "#/components/schemas/PanoramaNodeSummary"
        visited_at:
          type: string
          format: date-time
...
    CreateLearningSessionRequest:
      type: object
      additionalProperties: false
      required: [respondent_code]
      properties:
        respondent_code:
          type: string
          pattern: ^RSP-[A-Z0-9]{6}$
    LearningSession:
      type: object
      required: [id, respondent_code, phase, current_panorama_node, created_at, updated_at]
      properties:
        id:
          type: integer
        respondent_code:
          type: string
          pattern: ^RSP-[A-Z0-9]{6}$
        phase:
          type: string
          enum: [pretest, exploration, posttest, self_efficacy, completed]
        current_panorama_node:
          oneOf:
            - $ref: "#/components/schemas/PanoramaNodeSummary"
            - type: "null"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    LearningSessionCreated:
      allOf:
        - $ref: "#/components/schemas/LearningSession"
        - type: object
          required: [write_token]
          properties:
            write_token:
              type: string
              minLength: 64
              maxLength: 64
...
    CreateLearningSessionResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/LearningSessionCreated"
...
    PanoramaNodeSummary:
      type: object
      required: [id, name, slug, panorama_url]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        panorama_url:
          type: string
          format: uri
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true